+++ internal/characters/tighnari/asc.go
@@ -7,6 +7,18 @@ import (
 	"github.com/genshinsim/gcsim/pkg/modifier"
 )
 
+func (c *char) a1() {
+	m := make([]float64, attributes.EndStatType)
+	m[attributes.EM] = 50
+	c.AddStatMod(character.StatMod{
+		Base:         modifier.NewBase("tighnari-a1", 4*60),
+		AffectedStat: attributes.EM,
+		Amount: func() ([]float64, bool) {
+			return m, true
+		},
+	})
+}
+
 func (c *char) a4() {
 	m := make([]float64, attributes.EndStatType)
 	c.AddAttackMod(character.AttackMod{
